fix: bootstrap.ts
